REFACTOR whereLike test methods from QueryBuilderTest into QueryBuilderWhereLikeTest
